build(pyproject.toml): add dependency-groups for `uv`
At present, `project.optional-dependencies` with a section `[dev]`
is a section useable for `pip install .[dev]`, however not to
`uv install .[dev]`.  `uv` on the other hand manages sections of
`[dependency-groups]`.  The present partial copy-paste is to allow
either workflow to get, install, and use the optional requirements
of the project.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>